days/day-11-access-control-rbac/requirements.txt
# Day 11: Access Control - RBAC, Row-Level Security
# Production-ready access control systems

# Core Security Libraries
cryptography==41.0.7          # Encryption and secure operations
pyjwt==2.8.0                  # JWT token handling
passlib==1.7.4                # Password hashing
bcrypt==4.1.2                 # Secure password hashing

# Database Drivers
psycopg2-binary==2.9.9        # PostgreSQL driver for RLS
sqlalchemy==2.0.23            # ORM with security features
redis==5.0.1                  # Session and cache management

# Access Control Libraries
casbin==1.17.6                # Policy-based access control
flask-principal==0.4.0        # Identity and permission management
authlib==1.2.1                # OAuth and authentication

# Audit and Monitoring
structlog==23.2.0             # Structured logging
prometheus-client==0.19.0     # Metrics collection
sentry-sdk==1.38.0            # Error tracking

# Data Validation and Security
pydantic==2.5.0               # Data validation with security
marshmallow==3.20.1           # Schema validation
cerberus==1.3.5               # Data validation

# Testing and Development
pytest==7.4.3                 # Testing framework
pytest-cov==4.1.0             # Coverage reporting
faker==20.1.0                 # Test data generation
factory-boy==3.3.0            # Test object factories

# Utilities
python-dotenv==1.0.0          # Environment management
click==8.1.7                  # CLI interface
tabulate==0.9.0               # Table formatting
colorama==0.4.6               # Colored output

# Optional: Enterprise Features
# ldap3==2.9.1                # LDAP integration
# python-saml==1.15.0         # SAML authentication
# azure-identity==1.15.0      # Azure AD integration